requirements.txt
setuptools
wheel
python-dotenv
mysql-connector-python
tqdm